fix some tests, add back in perf tests, only have 1 test file for frequency
